// src/services/generated/apiTypes/taskResource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Laravel
 * OpenAPI spec version: 0.0.1
 */
import type { DocumentResource } from './documentResource';
import type { NoteResource } from './noteResource';
import type { TaskPriority } from './taskPriority';
import type { TaskStatus } from './taskStatus';
import type { UserResource } from './userResource';
import type { VisitResource } from './visitResource';

export interface TaskResource {
  id: number;
  visit_id: number;
  title: string;
  /** @nullable */
  description: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  due_date: string | null;
  /** @nullable */
  completed_at: string | null;
  assigned_by: number;
  /** @nullable */
  assigned_to: number | null;
  assigner?: UserResource;
  assignee?: UserResource | null;
  visit?: VisitResource;
  notes?: NoteResource[];
  documents?: DocumentResource[];
  /** @nullable */
  created_at: string | null;
  /** @nullable */
  updated_at: string | null;
}