// src/services/generated/apiTypes/visitStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Laravel
 * OpenAPI spec version: 0.0.1
 */

export type VisitStatus = typeof VisitStatus[keyof typeof VisitStatus];


export const VisitStatus = {
  scheduled: 'scheduled',
  completed: 'completed',
  cancelled: 'cancelled',
  no_show: 'no_show',
  rescheduled: 'rescheduled',
  pending: 'pending',
} as const;